renamed reducer, actions and selectors. added exports from index
