+++ packages/social-metadata-forms/src/index.js
@@ -1,1 +1,6 @@
+import * as actions from "./redux/actions/formActions";
+import * as reducers from "./redux/reducers/index";
+import * as selectors from "./redux/selectors";
+
 export { default as SocialMetadataPreviewForm } from "./SocialMetadataPreviewForm";
+export  { actions, reducers, selectors };
